+++ dashboards/customized/influxdb_v3/plugin/influx/types.py
@@ -1,0 +1,21 @@
+from typing import Any, Dict, Protocol, runtime_checkable
+
+
+@runtime_checkable
+class LineBuilderProtocol(Protocol):
+    def __init__(self, measurement: str) -> None: ...
+    def time_ns(self, value: int) -> None: ...
+    def tag(self, key: str, value: str) -> None: ...
+    def string_field(self, key: str, value: str) -> None: ...
+    def int64_field(self, key: str, value: int) -> None: ...
+    def float64_field(self, key: str, value: float) -> None: ...
+    def bool_field(self, key: str, value: bool) -> None: ...
+    def build(self) -> str: ...
+
+
+@runtime_checkable
+class InfluxDBClient(Protocol):
+    def query(self, query: str) -> list[Dict[str, Any]]: ...
+    def write_to_db(self, db_name: str, line: LineBuilderProtocol) -> None: ...
+    def info(self, msg: str) -> None: ...
+    def error(self, msg: str) -> None: ...
